+++ luminance/examples/02-render-state.rs
@@ -83,8 +83,7 @@ fn main() {
     .build()
     .unwrap();
 
-  let size = surface.size();
-  let mut back_buffer = Framebuffer::back_buffer(&mut surface, size);
+  let mut back_buffer = surface.back_buffer().unwrap();
 
   let mut blending = None;
   let mut depth_method = DepthMethod::Under;
@@ -116,8 +115,7 @@ fn main() {
     }
 
     if resize {
-      let size = surface.size();
-      back_buffer = Framebuffer::back_buffer(&mut surface, size);
+      back_buffer = surface.back_buffer().unwrap();
       resize = false;
     }