+++ luminance/examples/04-shader-uniforms.rs
@@ -62,8 +62,7 @@ fn main() {
     .build()
     .unwrap();
 
-  let size = surface.size();
-  let mut back_buffer = Framebuffer::back_buffer(&mut surface, size);
+  let mut back_buffer = surface.back_buffer().unwrap();
 
   // position of the triangle
   let mut triangle_pos = [0., 0.];
@@ -110,8 +109,7 @@ fn main() {
     }
 
     if resize {
-      let size = surface.size();
-      back_buffer = Framebuffer::back_buffer(&mut surface, size);
+      back_buffer = surface.back_buffer().unwrap();
       resize = false;
     }