+++ luminance/examples/08-shader-uniforms-adapt.rs
@@ -100,8 +100,7 @@ fn main() {
     .build()
     .unwrap();
 
-  let size = surface.size();
-  let mut back_buffer = Framebuffer::back_buffer(&mut surface, size);
+  let mut back_buffer = surface.back_buffer().unwrap();
   let mut triangle_pos = [0., 0.];
   let start_t = Instant::now();
   let mut resize = false;
@@ -148,8 +147,7 @@ fn main() {
     }
 
     if resize {
-      let size = surface.size();
-      back_buffer = Framebuffer::back_buffer(&mut surface, size);
+      back_buffer = surface.back_buffer().unwrap();
       resize = false;
     }